+++ CopilotKit/packages/react-ui/src/components/chat/messages/AssistantMessage.tsx
@@ -13,6 +13,7 @@ export const AssistantMessage = (props: AssistantMessageProps) => {
     onThumbsUp,
     onThumbsDown,
     isCurrentMessage,
+    feedback,
     markdownTagRenderers,
     showMessageControls = false,
   } = props;
@@ -37,11 +38,15 @@ export const AssistantMessage = (props: AssistantMessageProps) => {
   };
 
   const handleThumbsUp = () => {
-    if (onThumbsUp && message) onThumbsUp(message);
+    if (onThumbsUp && message) {
+      onThumbsUp(message);
+    }
   };
 
   const handleThumbsDown = () => {
-    if (onThumbsDown && message) onThumbsDown(message);
+    if (onThumbsDown && message) {
+      onThumbsDown(message);
+    }
   };
 
   const LoadingIcon = () => <span>{icons.activityIcon}</span>;
@@ -80,7 +85,9 @@ export const AssistantMessage = (props: AssistantMessageProps) => {
               </button>
               {onThumbsUp && (
                 <button
-                  className="copilotKitMessageControlButton"
+                  className={`copilotKitMessageControlButton ${
+                    feedback === "thumbsUp" ? "active" : ""
+                  }`}
                   onClick={handleThumbsUp}
                   aria-label={labels.thumbsUp}
                   title={labels.thumbsUp}
@@ -90,7 +97,9 @@ export const AssistantMessage = (props: AssistantMessageProps) => {
               )}
               {onThumbsDown && (
                 <button
-                  className="copilotKitMessageControlButton"
+                  className={`copilotKitMessageControlButton ${
+                    feedback === "thumbsDown" ? "active" : ""
+                  }`}
                   onClick={handleThumbsDown}
                   aria-label={labels.thumbsDown}
                   title={labels.thumbsDown}