+++ CopilotKit/packages/shared/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @copilotkit/shared
 
+## 1.10.6
+
+### Patch Changes
+
+- e0dd5d5: - feat: allow additional config in direct to llm actions
+
+## 1.10.6-next.6
+
+## 1.10.6-next.5
+
+### Patch Changes
+
+- e0dd5d5: - feat: allow additional config in direct to llm actions
+
+## 1.10.6-next.4
+
+## 1.10.6-next.3
+
+## 1.10.6-next.2
+
+## 1.10.6-next.1
+
+## 1.10.6-next.0
+
+## 1.10.5
+
+## 1.10.5-next.10
+
+## 1.10.5-next.9
+
+## 1.10.5-next.8
+
+## 1.10.5-next.7
+
+## 1.10.5-next.6
+
+## 1.10.5-next.5
+
+## 1.10.5-next.4
+
+## 1.10.5-next.3
+
+## 1.10.5-next.2
+
+## 1.10.5-next.1
+
+## 1.10.5-next.0
+
 ## 1.10.4
 
 ### Patch Changes